+++ package.json
@@ -12,7 +12,7 @@
     "react-dom": "^16.7.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1"
+    "react-scripts": "^2.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
