+++ public/index.html
@@ -20,7 +20,7 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <title>React App</title>
+    <title>RuDoc</title>
   </head>
   <body background="lol.png">
     <noscript>
